pyproject.toml
[tool.poetry]
name = "pyno"
version = "0.1.0"
description = ""
authors = ["7cm-diameter <from.this.ridiculous.loop@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
python-language-server = {extras = ["all"], version = "^0.36.2"}
isort = "^5.13.2"
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
